refactor: Cut out the path drawing part from the RenderDrawingOverLay component into a separate file and create hooks and utils.
