+++ src/components/camera/drawing/index.ts
@@ -1,0 +1,1 @@
+export { DrawingCanvas } from './DrawingCanvas';
